+++ seamonkeynovnc.sh
@@ -20,9 +20,18 @@ MimeType=text/html;text/xml;application/xhtml+xml;application/xml;application/rd
 Keywords=Browser;
 EOF
 sudo mv ~/seamonkey.desktop /usr/share/applications/
-tigervncserver  -SecurityTypes none  --I-KNOW-THIS-IS-INSECURE -xstartup /usr/bin/openbox -geometry 1366x768 -localhost no :0
+read -p "do you want to add a password to the novnc server? (yes/no) " yesorno
+if [[ "$yesorno" = "yes" ]]; then
+vncpasswd
+tigervncserver  -SecurityTypes none  --I-KNOW-THIS-IS-INSECURE  -xstartup /usr/bin/openbox -geometry 1366x768 -localhost no :0
 websockify -D --web=/usr/share/novnc/  --cert=~/linux-novnc/novnc.pem 6080 localhost:5900
-echo -e "seamonkey has started go to https://localhost:6080 to access your vnc session"
 sudo ln -s /usr/lib/seamonkey/seamonkey /usr/bin/startseamonkey
 export DISPLAY=:0
 startseamonkey
+elif [[ "$yesorno" = "no" ]]; then
+tigervncserver  -SecurityTypes none  --I-KNOW-THIS-IS-INSECURE -xstartup /usr/bin/openbox -geometry 1366x768 -localhost no :0
+websockify -D --web=/usr/share/novnc/  --cert=~/linux-novnc/novnc.pem 6080 localhost:5900
+sudo ln -s /usr/lib/seamonkey/seamonkey /usr/bin/startseamonkey
+export DISPLAY=:0
+startseamonkey
+else
